# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "409": {
            "description": "Email already registered"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in successfully"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Log in",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/routes": {
      "get": {
        "operationId": "RoutesController_listRoutes",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ARE_YENG",
                "TBS"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of routes"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all routes",
        "tags": [
          "Routes"
        ]
      }
    },
    "/api/routes/{id}": {
      "get": {
        "operationId": "RoutesController_getRouteDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route detail"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get route detail",
        "tags": [
          "Routes"
        ]
      }
    },
    "/api/routes/{id}/eta": {
      "get": {
        "operationId": "RoutesController_getRouteEta",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route ETA"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get route ETA with stops timeline",
        "tags": [
          "Routes"
        ]
      }
    },
    "/api/stops": {
      "get": {
        "operationId": "StopsController_listStops",
        "parameters": [
          {
            "name": "routeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of stops"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all stops",
        "tags": [
          "Stops"
        ]
      }
    },
    "/api/stops/nearby": {
      "get": {
        "operationId": "StopsController_getNearbyStops",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nearby stops"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get nearby stops",
        "tags": [
          "Stops"
        ]
      }
    },
    "/api/stops/{id}/eta": {
      "get": {
        "operationId": "StopsController_getStopEta",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stop ETA"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get ETAs for buses approaching a stop",
        "tags": [
          "Stops"
        ]
      }
    },
    "/api/buses/live": {
      "get": {
        "operationId": "BusesController_getLiveBuses",
        "parameters": [
          {
            "name": "routeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Live bus positions"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get live bus positions",
        "tags": [
          "Buses"
        ]
      }
    }
  },
  "info": {
    "title": "Tshwane Bus Tracker API",
    "description": "API for the Tshwane A Re Yeng Bus Tracking system",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "securePass123"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "securePass123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      }
    }
  }
}